+++ src/server/dev-server/index.js
@@ -1,23 +1,23 @@
-const cwd = process.cwd();
+// const cwd = process.cwd();
 
-require('@babel/register')({
-  extensions: ['.js'],
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-react"
-  ],
-  "plugins": [
-    "transform-es2015-modules-commonjs",
-    "@babel/plugin-proposal-class-properties"
-  ],
+// require('@babel/register')({
+//   extensions: ['.js'],
+//   "presets": [
+//     "@babel/preset-env",
+//     "@babel/preset-react"
+//   ],
+//   "plugins": [
+//     "transform-es2015-modules-commonjs",
+//     "@babel/plugin-proposal-class-properties"
+//   ],
 
-  ignore: [function (filename) {
+//   ignore: [function (filename) {
 
-    return filename.indexOf(cwd + `/node_modules/`) === 0;
-  }],
-});
+//     return filename.indexOf(cwd + `/node_modules/`) === 0;
+//   }],
+// });
 
-require("@babel/polyfill");
+// require("@babel/polyfill");
 
 
 const CoreModule = require("../../").CmsModule;